requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
Pillow==8.1.0
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
djangorestframework-simplejwt==4.6.0
